+++ example/android/app/src/main/kotlin/id/duende/example/MainActivity.kt
@@ -1,0 +1,6 @@
+package id.duende.example
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity: FlutterActivity() {
+}